Moved BmpParser to fastparse unit-tests and renamed byteparse module to classparse (#116)
* Added initial documentaion

* Minor changes in BmpParser, build.sbt and first iteration of new docs

* Second iteration of changing docs

* Moved BmpParser to fastparse unit-tests and renamed byteparse to the classparse

* Added classparse sources

* Fixes after rebase

* More renamings and fix travis config

* Minor Fix
