+++ classparse/jvm/src/test/scala/classparse/ProjectTests.scala
@@ -1,4 +1,3 @@
-package byteparse
 package classparse
 
 import java.io.File
