+++ classparse/shared/src/main/scala/classparse/ClassAttributes.scala
@@ -1,13 +1,12 @@
-package byteparse
 package classparse
 
 import fastparse.byte._
 
 object ClassAttributes {
-  import ClassParser.Info._
-  import ClassParser.BasicElems._
-  import ClassParser.Ast._
-  import ClassParser._
+  import ClassParse.Info._
+  import ClassParse.BasicElems._
+  import ClassParse.Ast._
+  import ClassParse._
   import CodeParser._
 
   sealed abstract class Attribute
