+++ classparse/shared/src/main/scala/classparse/ClassParse.scala
@@ -1,4 +1,3 @@
-package byteparse
 package classparse
 
 import fastparse.byte._
@@ -6,7 +5,7 @@ import fastparse.byte._
 import scala.collection.mutable.ArrayBuffer
 // https://docs.oracle.com/javase/specs/jvms/se7/html/jvms-4.html#jvms-4.7
 
-object ClassParser {
+object ClassParse {
 
   object BasicElems {
