+++ classparse/shared/src/test/resources/AttributeTest$InnerClass.class
