+++ classparse/shared/src/test/resources/AttributeTest.class
