+++ classparse/shared/src/test/resources/AttributeTest.java
