+++ classparse/shared/src/test/resources/Book.class
