+++ classparse/shared/src/test/resources/Book.java
