+++ classparse/shared/src/test/resources/Book2.java
