+++ fastparse/jvm/src/test/resources/lena.bmp
