+++ fastparse/jvm/src/test/resources/mirrored_lena.bmp
