+++ fastparse/jvm/src/test/scala/fastparse/LargeBmpIteratorTests.scala
@@ -1,14 +1,13 @@
-package byteparse
+package fastparse
 
 import java.io.InputStream
 
 import fastparse.{IteratorParserInput, Utils}
 import utest._
 import fastparse.byte._
-import BmpParser._
+import fastparse.BmpTests.BmpParse._
 
 import scala.collection.mutable
-
 object LargeBmpIteratorTests extends TestSuite {
 
   class StreamToIteratorByte(stream: InputStream, bufferSize: Int) extends Iterator[Array[Byte]] {
@@ -41,7 +40,7 @@ object LargeBmpIteratorTests extends TestSuite {
 
   val tests = TestSuite {
     'large {
-      val Parsed.Success(_, i) = bmp.parseIterator(lenaIterator)
+      val Parsed.Success(_, i) = BmpTests.BmpParse.bmp.parseIterator(lenaIterator)
       val expectedIndex = lenaIterator.map(_.length).sum
       assert(i == expectedIndex)
     }
@@ -56,7 +55,7 @@ object LargeBmpIteratorTests extends TestSuite {
         }
       }
 
-      val Parsed.Success(_, i) = bmp.parseInput(loggedInput)
+      val Parsed.Success(_, i) = BmpTests.BmpParse.bmp.parseInput(loggedInput)
       println(s"Size: ${lenaIterator.map(_.length).sum}")
       println(s"Max buffer length: ${loggedInput.maxInnerLength}")
     }
