+++ perftests/jvm/src/test/scala/perftests/binary/BmpParse.scala
@@ -6,11 +6,11 @@ import perftests.Utils
 import utest._
 
 
-object ByteParse extends TestSuite {
+object BmpParse extends TestSuite {
   val lenaRecource = getClass.getResource("/lena.bmp")
   val lenaSource = Files.readAllBytes(Paths.get(lenaRecource.toURI.getPath))
   def lenaIterator(size: Int) = lenaSource.grouped(size)
-  val parser = byteparse.BmpParser.bmp
+  val parser = fastparse.BmpTests.BmpParse.bmp
 
   val tests = TestSuite {
     'Lena {
