+++ perftests/jvm/src/test/scala/perftests/binary/ClassParse.scala
@@ -2,7 +2,6 @@ package perftests.binary
 
 import java.nio.file.{Files, Paths}
 
-import byteparse.classparse.ClassParser
 import perftests.Utils
 import utest._
 
@@ -11,7 +10,7 @@ object ClassParse extends TestSuite {
   val collisionResource = getClass.getResource("/CollisionJNI.class")
   val collisionSource = Files.readAllBytes(Paths.get(collisionResource.toURI.getPath))
   def collisionIterator(size: Int) = collisionSource.grouped(size)
-  val parser = ClassParser.classFile
+  val parser = classparse.ClassParse.classFile
 
   val tests = TestSuite {
     Utils.benchmarkAll(
